Add hash table resize method
Add hash table resize method, to be checked. It is able to automatically
resize the hash table given the nodes number to be stored computing
a well balanced prime number, closeset to the query, by means of the
Sieve of Eratosthenes.

Minor changes in initialize and destroy methods too.

A testing facility must be added, probably with a new key format to be
submitted to @francescosalvadore approval.