Add finest code computation and lower/greater comparison
Add finest code computation and lower/greater comparison: finest code of
a queried code is computed as the latest child in the maximum refinement
level allowed for the tree. The finest code is used into the comparison
of two codes: now I am ready for the tree sorting for the MPI balancing.

@francescosalvadore may have a good sorting algorithm in mind, ask him

For @andreadimascio : hash table has been saved, Francesco creates
a good simple looping method.