Clean namespaces for NVFortran stupidity
Clean namespaces for NVFortran stupidity: rename all variables
that confuse NVFortran in tree object.

Why:

NVFortran is not able (sometime) to resolve correctly the use of
variables having the same name of module functions, but this happens
randomly, often it is able to do it, sometimes not...

Side effects:

Waste of hours.